Feat: Implemented UI Filter
