+++ foodbook_app/lib/bloc/browse_bloc/browse_event.dart
@@ -15,3 +15,9 @@ class FilterRestaurants extends BrowseEvent {
   });
 }
 
+class ToggleBookmark extends BrowseEvent {
+  final String restaurantId;
+
+  ToggleBookmark(this.restaurantId);
+}
+
