+++ foodbook_app/lib/presentation/views/restaurant_views/browse_view.dart
@@ -1,28 +1,18 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:foodbook_app/bloc/browse_bloc/browse_bloc.dart';
-import 'package:foodbook_app/bloc/browse_bloc/browse_event.dart';
-import 'package:foodbook_app/data/models/restaurant.dart';
-import 'package:foodbook_app/data/repository/restaurant_repo.dart';
-import 'package:foodbook_app/presentation/widgets/restaurant_card.dart';
+import 'package:foodbook_app/bloc/browse_bloc/browse_state.dart';
+import 'package:foodbook_app/presentation/widgets/menu/navigation_bar.dart';
+import 'package:foodbook_app/presentation/widgets/menu/filter_bar.dart';
+import 'package:foodbook_app/presentation/widgets/restaurant_card/restaurant_card.dart';
 
-// ... Other imports remain unchanged
+// Asegúrate de tener todos los imports necesarios aquí
 
-class BrowseView extends StatefulWidget {
+class BrowseView extends StatelessWidget {
   BrowseView({Key? key}) : super(key: key);
 
   @override
-  _BrowseViewState createState() => _BrowseViewState();
-}
-
-class _BrowseViewState extends State<BrowseView> {
-  final RestaurantRepository repository = RestaurantRepository();
-  int _selectedIndex = 0;
-
-  @override
   Widget build(BuildContext context) {
-    List<Restaurant> restaurants = repository.fetchRestaurants();
-
     return Scaffold(
       appBar: AppBar(
         backgroundColor: Colors.white, // Set AppBar background to white
@@ -33,12 +23,9 @@ class _BrowseViewState extends State<BrowseView> {
             color: Colors.black, // Title color
           ),
         ),
-        iconTheme: IconThemeData(color: Colors.black), // Appbar icons color
-        actions: <Widget>[
-          IconButton(
-            icon: Icon(Icons.filter_list, color: Colors.black), // Icon for filter
-            onPressed: _showFilterDialog, // Show the filter dialog when the icon is pressed
-          ),
+        // Add the FilterBar widget to the AppBar
+        actions: [
+          FilterBar(),
         ],
         elevation: 0, // Remove shadow
       ),
@@ -71,171 +58,29 @@ class _BrowseViewState extends State<BrowseView> {
             color: Colors.grey[300], // Color of the divider line
           ),
           Expanded(
-            child: ListView.builder(
-              itemCount: restaurants.length,
-              itemBuilder: (context, index) {
-                return RestaurantCard(restaurant: restaurants[index]);
+            child: BlocBuilder<BrowseBloc, BrowseState>(
+              builder: (context, state) {
+                if (state is RestaurantsLoadInProgress) {
+                  return Center(child: CircularProgressIndicator());
+                } else if (state is RestaurantsLoadSuccess) {
+                  return ListView.builder(
+                    itemCount: state.restaurants.length,
+                    itemBuilder: (context, index) {
+                      return RestaurantCard(restaurant: state.restaurants[index]);
+                    },
+                  );
+                } else if (state is RestaurantsLoadFailure) {
+                  return Center(child: Text('Failed to load restaurants'));
+                }
+                // Si el estado inicial es RestaurantsInitial o cualquier otro estado no esperado
+                return Center(child: Text('Start browsing by applying some filters!'));
               },
             ),
           ),
         ],
       ),
-      bottomNavigationBar: BottomNavigationBar(
-        items: <BottomNavigationBarItem>[
-          BottomNavigationBarItem(
-            icon: Icon(Icons.search),
-            label: 'Browse',
-          ),
-          BottomNavigationBarItem(
-            icon: Icon(Icons.star_border),
-            label: 'For you',
-          ),
-          BottomNavigationBarItem(
-            icon: Icon(Icons.bookmark_border),
-            label: 'Bookmarks',
-          ),
-        ],
-        currentIndex: _selectedIndex,
-        selectedItemColor: Colors.blue,
-        unselectedItemColor: Colors.grey,
-        onTap: (index) {
-          setState(() {
-            _selectedIndex = index;
-          });
-          // Add navigation logic here if necessary
-        },
-      ),
+      bottomNavigationBar: CustomNavigationBar(),
     );
   }
-
-
-  void _showFilterDialog() {
-    String selectedPrice = '\$';
-    double selectedDistance = 5.0;
-    String selectedCategory = '';
-    bool isPriceFilterEnabled = false;
-    bool isDistanceFilterEnabled = false;
-    bool isCategoryFilterEnabled = false;
-
-    showModalBottomSheet(
-      context: context,
-      builder: (context) {
-        return StatefulBuilder(
-          builder: (BuildContext context, StateSetter setState) {
-            return Container(
-              padding: EdgeInsets.all(16),
-              child: Wrap(
-                children: <Widget>[
-                  ListTile(
-                    title: Text('Price Range'),
-                    trailing: Switch(
-                      value: isPriceFilterEnabled,
-                      onChanged: (bool value) {
-                        setState(() {
-                          isPriceFilterEnabled = value;
-                        });
-                      },
-                    ),
-                    onTap: () {
-                      setState(() {
-                        isPriceFilterEnabled = !isPriceFilterEnabled;
-                      });
-                    },
-                  ),
-                  isPriceFilterEnabled
-                      ? DropdownButton<String>(
-                          value: selectedPrice,
-                          onChanged: (value) {
-                            setState(() {
-                              selectedPrice = value!;
-                            });
-                          },
-                          items: ['\$', '\$\$', '\$\$\$', '\$\$\$\$']
-                              .map<DropdownMenuItem<String>>((String value) {
-                            return DropdownMenuItem<String>(
-                              value: value,
-                              child: Text(value),
-                            );
-                          }).toList(),
-                        )
-                      : Container(),
-                  ListTile(
-                    title: Text('Distance (km)'),
-                    trailing: Switch(
-                      value: isDistanceFilterEnabled,
-                      onChanged: (bool value) {
-                        setState(() {
-                          isDistanceFilterEnabled = value;
-                        });
-                      },
-                    ),
-                    onTap: () {
-                      setState(() {
-                        isDistanceFilterEnabled = !isDistanceFilterEnabled;
-                      });
-                    },
-                  ),
-                  isDistanceFilterEnabled
-                      ? Slider(
-                          min: 0.0,
-                          max: 20.0,
-                          divisions: 20,
-                          label: '${selectedDistance.toStringAsFixed(1)} km',
-                          value: selectedDistance,
-                          onChanged: (value) {
-                            setState(() {
-                              selectedDistance = value;
-                            });
-                          },
-                        )
-                      : Container(),
-                  ListTile(
-                    title: Text('Category'),
-                    trailing: Switch(
-                      value: isCategoryFilterEnabled,
-                      onChanged: (bool value) {
-                        setState(() {
-                          isCategoryFilterEnabled = value;
-                        });
-                      },
-                    ),
-                    onTap: () {
-                      setState(() {
-                        isCategoryFilterEnabled = !isCategoryFilterEnabled;
-                      });
-                    },
-                  ),
-                  isCategoryFilterEnabled
-                      ? TextField(
-                          decoration: InputDecoration(
-                            labelText: 'Category',
-                            hintText: 'e.g., Italian, Vegan',
-                          ),
-                          onChanged: (value) {
-                            selectedCategory = value;
-                          },
-                        )
-                      : Container(),
-                  ElevatedButton(
-                    onPressed: () {
-                      context.read<BrowseBloc>().add(FilterRestaurants(
-                        price: isPriceFilterEnabled ? selectedPrice : null,
-                        distance: isDistanceFilterEnabled ? selectedDistance : null,
-                        category: isCategoryFilterEnabled ? selectedCategory : null,
-                      ));
-                      Navigator.pop(context);
-                    },
-                    child: Text('Apply Filter'),
-                  ),
-                ],
-              ),
-            );
-          },
-        );
-      },
-    );
-  }
-
-
-
 }
+
