+++ foodbook_app/lib/presentation/widgets/menu/navigation_bar.dart
@@ -1,0 +1,41 @@
+import 'package:flutter/material.dart';
+
+class CustomNavigationBar extends StatefulWidget {
+  @override
+  _CustomNavigationBarState createState() => _CustomNavigationBarState();
+}
+
+class _CustomNavigationBarState extends State<CustomNavigationBar> {
+  int _selectedIndex = 0; // Índice inicial seleccionado
+
+  void _onItemTapped(int index) {
+    setState(() {
+      _selectedIndex = index;
+    });
+    // Aquí puedes agregar tu lógica de navegación si es necesario
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return BottomNavigationBar(
+      items: const <BottomNavigationBarItem>[
+        BottomNavigationBarItem(
+          icon: Icon(Icons.search),
+          label: 'Browse',
+        ),
+        BottomNavigationBarItem(
+          icon: Icon(Icons.star_border),
+          label: 'For You',
+        ),
+        BottomNavigationBarItem(
+          icon: Icon(Icons.bookmark_border),
+          label: 'Bookmarks',
+        ),
+      ],
+      currentIndex: _selectedIndex, // Índice actual basado en la selección del usuario
+      selectedItemColor: Colors.blue,
+      unselectedItemColor: Colors.grey,
+      onTap: _onItemTapped, // Método que se llama cuando se selecciona un ítem
+    );
+  }
+}
