+++ foodbook_app/lib/presentation/widgets/restaurant_card/subtitle_section.dart
@@ -1,0 +1,34 @@
+import 'package:flutter/material.dart';
+import 'package:foodbook_app/data/models/restaurant.dart';
+
+class SubtitleSection extends StatelessWidget {
+  final Restaurant restaurant;
+
+  const SubtitleSection({Key? key, required this.restaurant}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      children: [
+        Icon(Icons.access_time, size: 16, color: Color.fromARGB(255, 0, 0, 0)),
+        SizedBox(width: 4),
+        Text(
+          '${restaurant.timeRange} · ',
+          style: TextStyle(
+            fontSize: 14,
+            color: Color.fromARGB(255, 53, 53, 53),
+          ),
+        ),
+        Icon(Icons.location_on, size: 16, color: Color.fromARGB(255, 0, 0, 0)),
+        SizedBox(width: 4),
+        Text(
+          '${restaurant.distance.toStringAsFixed(1)} km',
+          style: TextStyle(
+            fontSize: 14,
+            color: Color.fromARGB(255, 53, 53, 53),
+          ),
+        ),
+      ],
+    );
+  }
+}
